Added in a small database to test what I had.
LetterGenerator:
I switched from using 3 lists to store the letter combos to a hashmap which allows us to access all of them.

Main:
I added in a sample run that you guys can edit to try different numbers.

UI:

Added in a very simple print function.

WordDatabase:

This is where most of my changes are in. I added in a constructor that reads the database and stored the valid words in a hashmap. The 3 letter prefix of a word will be the key and the value will be a list of words that start with that prefix

There is also a function that allows us to query the database to determine what letter combinations are a valid word. It will then return a list of all the valid words.