I added in descriptions to each class to outline their purpose.
Classes I added:

AutomatedTesting: Gonna add logic to test different components of the program

Relevance Generator: This is gonna help us manage the words the user is liking so we can eventually Display their top liked words.

Changes:

UI class:

I added in a small javaFX script so I could simulate getting user input.
Also wanted to get an idea of how we would be combining the UI and the rest of the program.

    Expected output:

    a small screen that asks for a phone number and a submit button
    once you enter a valid number and hit submit then it should display
    a table of valid words