I added more words into the database to test the number 4096692255 I also made attributes in the Phone and Word class private. Because of this I also added in getterMethods for those attributes.
In the UI I added in functionality so we can push the enter key to press the buttons instead of having to click on them directly.

Most of the work I did was in the UI for the Display Results screen. I added in functionality which cycles through the word list and allows the user to select and narrow down their results.

*also renamed an image file